# Conflicts:
#	server-user/src/main/java/com/onek/user/StoreManageModule.java
#	server-user/src/main/java/com/onek/user/operations/UpdateAuditOp.kt

+++ application-global/src/main/java/com/onek/calculate/filter/QualFilter.java
@@ -40,7 +40,7 @@ public class QualFilter extends BaseFilter {
             Activity act = (Activity) activity;
 
             int qualCode = act.getQualcode();
-            int qualValue = act.getQualvalue();
+            long qualValue = act.getQualvalue();
 
             switch (qualCode) {
                 case ALL:
